move rowToString and stringStringToRow to FeatureListUtils
