+++ mzmine-community/src/main/java/io/github/mzmine/modules/dataprocessing/filter_deleterows/DeleteRowsParameters.java
@@ -32,7 +32,7 @@ import io.github.mzmine.parameters.impl.SimpleParameterSet;
 import io.github.mzmine.parameters.parametertypes.StringParameter;
 import io.github.mzmine.parameters.parametertypes.selectors.FeatureListsParameter;
 import io.github.mzmine.parameters.parametertypes.selectors.FeatureListsSelection;
-import io.github.mzmine.util.FeatureUtils;
+import io.github.mzmine.util.FeatureListUtils;
 import java.util.List;
 
 public class DeleteRowsParameters extends SimpleParameterSet {
@@ -52,7 +52,7 @@ public class DeleteRowsParameters extends SimpleParameterSet {
 
     parameterSet.setParameter(DeleteRowsParameters.flist,
         new FeatureListsSelection((ModularFeatureList) featureList));
-    parameterSet.setParameter(rowIds, FeatureUtils.rowsToIdString(rows));
+    parameterSet.setParameter(rowIds, FeatureListUtils.rowsToIdString(rows));
 
     return parameterSet;
   }
