+++ mzmine-community/src/main/java/io/github/mzmine/util/FeatureUtils.java
@@ -60,7 +60,6 @@ import io.github.mzmine.util.spectraldb.entry.SpectralDBAnnotation;
 import java.text.Format;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -662,15 +661,4 @@ public class FeatureUtils {
     return absCharge * pol.getSign();
   }
 
-  public static String rowsToIdString(List<? extends FeatureListRow> rows) {
-    return rows.stream().map(FeatureListRow::getID).map(Object::toString)
-        .collect(Collectors.joining(";"));
-  }
-
-  public static List<FeatureListRow> idStringToRows(ModularFeatureList flist, String str) {
-    final Set<Integer> ids = Arrays.stream(str.split(";")).map(Integer::valueOf)
-        .collect(Collectors.toSet());
-    return flist.stream().filter(row -> ids.contains(row.getID()))
-        .sorted(Comparator.comparingInt(FeatureListRow::getID)).toList();
-  }
 }
